+++ src/main/groovy/com/devbliss/docker/DockerPlugin.groovy
@@ -1,6 +1,8 @@
 package com.devbliss.docker
 
 import com.devbliss.docker.task.BuildAndPushDockerImageTask
+import com.devbliss.docker.task.CleanupOldContainersTask
+import com.devbliss.docker.task.PullDependencyImages
 import com.devbliss.docker.task.GetServiceDependenciesTask
 import com.devbliss.docker.task.StartDependenciesTask
 import com.devbliss.docker.task.StopAllRunningContainersTask
@@ -13,8 +15,6 @@ import org.slf4j.Logger
 import org.slf4j.LoggerFactory
 
 /**
- * @author Christian Soth <christian.soth@devbliss.com> on 12.01.15.
- *
  * Devbliss Docker plugin to create a set of default docker tasks for a project.
  * The plugin extends the DockerPlugin from gesellix with additional tasks and configuration for devbliss needs.
  */
@@ -24,9 +24,9 @@ class DockerPlugin implements Plugin<Project> {
     public void apply(Project project) {
         project.getPlugins().apply(ParentDockerPlugin)
 
-        project.task("startDependencies", type: StartDependenciesTask)
+        project.task(Configuration.TASK_NAME_START_DEPENDENCIES, type: StartDependenciesTask)
         project.task("stopAllRunningContainers", type: StopAllRunningContainersTask)
-        project.task("serviceDependencies", type: GetServiceDependenciesTask)
+        project.task(Configuration.TASK_NAME_GET_SERVICE_DEPENDENCIES, type: GetServiceDependenciesTask)
         BuildAndPushDockerImageTask buildAndPushDockerImage = project.task('buildAndPushDockerImage', type: BuildAndPushDockerImageTask)
         DockerPullTask dockerPullTask = project.task('pullDockerImage', type: DockerPullTask)
         DockerPushTask dockerPushTask = project.task('pushDockerImage', type: DockerPushTask)
@@ -35,7 +35,9 @@ class DockerPlugin implements Plugin<Project> {
         DockerRunTask dockerRunTask = project.task('runDockerContainer', type: DockerRunTask)
         DockerRmTask dockerRmTask = project.task('removeDockerContainer', type: DockerRmTask)
         DockerBuildTask dockerBuildTask = project.task('buildDockerImage', type: DockerBuildTask)
-
+        PullDependencyImages pullDependencyImages = project.task('pullDependencyImages', type: PullDependencyImages)
+        CleanupOldContainersTask cleanupOldContainersTask = project.task('cleanupOldContainers', type: CleanupOldContainersTask)
+    
         Configuration configuration = new Configuration(project)
 
         //Tasks that depend on other tasks