+++ src/test/groovy/com/devbliss/docker/util/TestTask.groovy
@@ -5,10 +5,6 @@ import org.gradle.api.tasks.Input
 import org.gradle.api.tasks.Optional
 import org.gradle.api.tasks.TaskAction
 
-/**
- * Created by Christian Soth <christian.soth@devbliss.com> on 14.01.15.
- */
-
 class TestTask extends AbstractDockerTask {
 
     @Input